% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack.R
\name{jack}
\alias{jack}
\title{Non-parametric Bootstrapping}
\usage{
jack(data, func = NULL)
}
\arguments{
\item{data}{he data as a vector}

\item{func}{function to be bootstrapped}
}
\value{
the standard error to be estimated
}
\description{
Compute the jackknife estimate of standard error using R
}
\examples{
\dontrun{
data(aircondit,package = "boot")
air <- aircondit$hours
jack(data = air, func = mean)
}
}
